Revert moving of transition label provider
