+++ de.dlr.sc.virsat.model.extension.statemachines.ui/src/de/dlr/sc/virsat/model/extension/statemachines/ui/diagram/features/transitions/TransitionAddFeature.java
@@ -15,7 +15,6 @@ import de.dlr.sc.virsat.model.dvlm.categories.CategoryAssignment;
 import de.dlr.sc.virsat.model.extension.statemachines.model.State;
 import de.dlr.sc.virsat.model.extension.statemachines.model.Transition;
 import de.dlr.sc.virsat.model.extension.statemachines.ui.diagram.features.abstractConnection.AbstractConnectionAddFeature;
-import de.dlr.sc.virsat.model.extension.statemachines.ui.diagram.features.label.LabelProviderInstantiator;
 /**
 * Feature for adding transition in an state machine diagram. 
 * */
